First commit - User registration and login functionality
User registration and login functionality for general user, doctor, hospital or lab and insurance companies. Storage of user object in IPFS and creation of key. Storage of key in js arrays and local storage to be upgraded later to Hyperledger Fabric. UI for file upload in the General user screen after login.